+++ src/Controller/ModulesController.php
@@ -62,7 +62,7 @@ class ModulesController extends AppController
         parent::beforeFilter($event);
 
         $actions = [
-            'export', 'delete', 'changeStatus',
+            'delete', 'changeStatus',
         ];
 
         if (in_array($this->request->params['action'], $actions)) {
@@ -367,11 +367,11 @@ class ModulesController extends AppController
     }
 
     /**
-     * Export data in csv format
+     * Bulk change status for objects
      *
-     * @return \Cake\Http\Response
+     * @return \Cake\Http\Response|null
      */
-    public function changeStatus() : Response
+    public function changeStatus() : ?Response
     {
         $this->request->allowMethod(['post']);
         $ids = $this->request->getData('ids');
@@ -388,81 +388,5 @@ class ModulesController extends AppController
         }
 
         return $this->redirect(['_name' => 'modules:list', 'object_type' => $this->objectType]);
-    }
-
-    /**
-     * Export data in csv format
-     *
-     * @return void
-     */
-    public function export() : void
-    {
-        $this->request->allowMethod(['post']);
-        $fields = ['id', 'title', 'description', 'uname', 'status', 'body', 'lang', 'extra'];
-        $delimiter = ',';
-        $ids = $this->request->getData('ids');
-        $csv = implode($delimiter, $fields) . "\n";
-        if (!empty($ids)) { // export selected (filter by id)
-            $response = $this->apiClient->getObjects($this->objectType, ['filter' => ['id' => $ids]]);
-            $csv .= $this->csvRows($response, $fields, $delimiter);
-            $csv = rtrim($csv, "\n"); // remove last newline
-            $this->set('csv', $csv);
-        } else { // export all
-            $response = $this->apiClient->getObjects($this->objectType, [
-                'page' => 1,
-                'page_size' => 100,
-            ]);
-            while (!empty($response['links']['next'])) {
-                $csv .= $this->csvRows($response, $fields, $delimiter);
-                $response = $this->apiClient->get($response['links']['next']);
-            }
-            $csv .= $this->csvRows($response, $fields, $delimiter);
-            $csv = rtrim($csv, "\n"); // remove last newline
-        }
-
-        // Output csv file
-        header("Content-type: text/csv");
-        header("Content-Disposition: attachment; filename=" . $this->objectType . "_" . date('Ymd-His') . ".csv");
-        header('Content-Transfer-Encoding: binary');
-        header('Expires: 0');
-        header('Cache-Control: must-revalidate, post-check=0, pre-check=0');
-        header("Pragma: no-cache");
-        print $csv;
-        exit;
-    }
-
-    /**
-     * Csv rows per response data
-     *
-     * @param array $response the Response
-     * @param array $fields the Fields to export for object
-     * @param string $delimiter the Delimiter char
-     * @return string The csv data
-     */
-    private function csvRows($response = [], $fields = [], $delimiter = ',') : string
-    {
-        if (empty($response) || empty($response['data'])) {
-            return '';
-        }
-        $csv = '';
-        foreach ($response['data'] as $index => $data) {
-            $cells = [];
-            foreach ($fields as $field) {
-                if (!empty($data[$field])) {
-                    $cells[] = '"' . preg_replace('/"/', '""', $data[$field]) . '"';
-                } elseif (!empty($data['attributes'][$field])) {
-                    if (is_array($data['attributes'][$field])) {
-                        $cells[] = '"' . preg_replace('/"/', '""', json_encode($data['attributes'][$field])) . '"';
-                    } else {
-                        $cells[] = '"' . preg_replace('/"/', '""', $data['attributes'][$field]) . '"';
-                    }
-                } else {
-                    $cells[] = '';
-                }
-            }
-            $csv .= implode($delimiter, $cells) . "\n";
-        }
-
-        return $csv;
     }
 }
